Create interactive game UI and rework code.
Added new html elements and javascript functions to make the game
playable using buttons, rather than the previous text prompts.
Now, the score changes show up along with messages as the buttons are
clicked. I also added a play again function, which allows the user to
click yes or no, resulting in the reset of the score at the end of best
of 5.